refactor(ui): split Avatar component and rename Buttons file
- 將 Avatar 元件拆分為 index 和 styled，提升可讀性與維護性
- 修正 Buttons.tsx 檔名為 buttons.tsx，符合匯出命名規範
