+++ components/ui/Link.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+import NextLink from "next/link";
+// export const BaseLink = styled(NextLink)`
+export const Link = styled(NextLink)`
+`;
+
+// TBD: 盡可能用兩個詞組成
