+++ types/uiProps.ts
@@ -1,0 +1,4 @@
+// TBD: 命名要改為只有使用 $isDefault 的 input 元件
+export type InputField = {
+  $isDefault?: boolean;
+};
