+++ utils/react-icons/IconColor.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { FaLine, FaFacebookSquare } from "react-icons/fa";
 
-export const IconWrapper = styled.div`
+export const IconWhiteWrapper = styled.div`
   display: inline-flex;
   align-items: center;
   justify-content: center;
@@ -16,12 +16,6 @@ export const StyledFaLine = styled(FaLine)`
   vertical-align: top;
   fill: #06c755;
 `;
-
-// export const StyledFaFacebook = styled(FaFacebookSquare)`
-//   vertical-align: top;
-//   fill: #3b579d;
-//   background-color: white;
-// `;
 
 export const StyledFaFacebook = styled(FaFacebookSquare)`
   fill: #3b579d;
